+++ completions/tspin.zsh
@@ -18,19 +18,14 @@ _tspin() {
 '--config-path=[Provide a custom path to a configuration file]:CONFIG_PATH:_files' \
 '(-f --follow)-c+[Capture the output (stdout) of a command and view it in \`less\`]:LISTEN_COMMAND:_default' \
 '(-f --follow)--listen-command=[Capture the output (stdout) of a command and view it in \`less\`]:LISTEN_COMMAND:_default' \
-'*--words-red=[Highlight the provided words in red]:WORDS_RED:_default' \
-'*--words-green=[Highlight the provided words in green]:WORDS_GREEN:_default' \
-'*--words-yellow=[Highlight the provided words in yellow]:WORDS_YELLOW:_default' \
-'*--words-blue=[Highlight the provided words in blue]:WORDS_BLUE:_default' \
-'*--words-magenta=[Highlight the provided words in magenta]:WORDS_MAGENTA:_default' \
-'*--words-cyan=[Highlight the provided words in cyan]:WORDS_CYAN:_default' \
+'*--highlight=[Highlights in the form color\:word1,word2 \[possible values\: red, green, yellow, blue, magenta, cyan\]]:COLOR_WORD:_default' \
 '*--enable=[Enable specific highlighters]:ENABLED_HIGHLIGHTERS:(numbers urls pointers dates paths quotes key-value-pairs uuids ip-addresses processes json)' \
 '*--disable=[Disable specific highlighters]:DISABLED_HIGHLIGHTERS:(numbers urls pointers dates paths quotes key-value-pairs uuids ip-addresses processes json)' \
 '-f[Follow the contents of a file]' \
 '--follow[Follow the contents of a file]' \
 '-p[Print the output to stdout]' \
 '--print[Print the output to stdout]' \
-'--no-builtin-keywords[Disable the highlighting of all builtin keyword groups (booleans, nulls, log severities and common REST verbs)]' \
+'--disable-builtin-keywords[Disable the highlighting of all builtin keyword groups (booleans, nulls, log severities and common REST verbs)]' \
 '--generate-bash-completions[Print bash completions to stdout]' \
 '--generate-fish-completions[Print fish completions to stdout]' \
 '--generate-zsh-completions[Print zsh completions to stdout]' \